+++ server/src/core/services/competition.ts
